backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
pydantic==2.4.2
python-dotenv==0.21.1
dashscope==1.16.0
aiosqlite==0.19.0
typing-extensions>=4.5
